+++ calculator/src/main/java/com/louis993546/calculator/CalculatorApp.kt
@@ -1,14 +1,24 @@
 package com.louis993546.calculator
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.*
-import androidx.compose.runtime.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.heightIn
+import androidx.compose.foundation.layout.padding
+import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.louis993546.metro.*
+import com.louis993546.metro.LocalAccentColor
+import com.louis993546.metro.LocalButtonColor
+import com.louis993546.metro.LocalTextOnAccentColor
+import com.louis993546.metro.LocalTextOnButtonColor
+import com.louis993546.metro.Text
 
 @Composable
 fun CalculatorApp(
