+++ demo/src/main/java/com/louis993546/metro/demo/MainActivity.kt
@@ -8,7 +8,15 @@ import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.aspectRatio
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -27,7 +35,10 @@ import com.google.accompanist.pager.ExperimentalPagerApi
 import com.google.accompanist.pager.HorizontalPager
 import com.google.accompanist.pager.rememberPagerState
 import com.louis993546.calculator.CalculatorApp
-import com.louis993546.metro.*
+import com.louis993546.metro.LocalAccentColor
+import com.louis993546.metro.LocalBackgroundColor
+import com.louis993546.metro.LocalTextOnAccentColor
+import com.louis993546.metro.Text
 import com.louis993546.metro.demo.theme.MetroDemoTheme
 import com.louis993546.metro_settings.MetroSettingsApp
 import timber.log.Timber
