+++ demo/src/test/java/com/louis993546/metro/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
 package com.louis993546.metro
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
