+++ metro/build.gradle
@@ -33,6 +33,7 @@ dependencies {
     api "androidx.compose.ui:ui:$compose_version"
     api "androidx.compose.foundation:foundation:$compose_version"
     api "androidx.compose.ui:ui-tooling:$compose_version"
+    api "com.google.accompanist:accompanist-pager:$accompanist_version"
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
