+++ metro_settings/build.gradle
@@ -32,7 +32,6 @@ android {
 
 dependencies {
     implementation project(path: ":metro")
-    implementation "com.google.accompanist:accompanist-pager:$accompanist_version"
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
